requirements.txt
pikudhaoref.py==0.0.6
pyrogram
tgcrypto
